Autoreleasing Memory
- Essentially, objects in memory that no one uses will be removed from
memory at one point. Kind of like garbage collection, but different.
When a variable in an autorelease pool is released, it’s tied to that
autorelease pool. When in turn the pool itself is released, all memory
in the pool is released too. It’s like batching the release of memory.

- Creating our own autorelease pool we’re helping the iPhone system
manage memory. Since we could be working with hundreds of species
objects within the autorelease pool, the memory could clog up with
undiscarded memory. The earliest point in time where the normal
autorelease pool discards memory is at the end of the method! Thus, we
run the risk of running out of memory because the autorelease mechanism
doesn’t discard quickly enough. By creating your own autorelease pool,
you can influence the discarding of released memory and avoid being
stuck for free memory.